Store Page Update
I got a "categories" bar and made it so the purchase table is centered
in the remaining space of the page.  Started messing with javascript to
detect button presses (category buttons pop up an alert when they're
pressed right now for testing).  Hopefully we can dynamically display
which purchase table is being shown based off the button click.