Deleted obsolete HTML pages, and made DDRC's logo not link to their site on other pages
I deleted homepage.html and login_page.html, since they were the first
prototypes and we were not using them. We have been using
login_page1.1.0 and homepage 2.0.1 instead (and I just now renamed those
to 1.1.0 and 2.1.1, respectively). Also, I went and changed all the
pages so that clicking on DDRC's logo does not go to their website, as
requested by the client. I will be updating the HTML files in the django
folder at a later time.