HTML pages in Django folder updated to not go to DDRC's website
I previously updated all the HTML pages in the HTML folder to not link
to DDRC's website when the logo was clicked, so I did the same for the
Django folder. I had to revert it because I forgot to sync the folder,
so I was working on an older version of it.