Store Page Functional, and Competition Form*
Formatted the store page so that it's consistent with the other pages.
Got it functioning correctly through Django and included a button back
to the index page, not too sure about the location of it though.  It's
images for the "User Pictures" table aren't static just because I don't
want to download the images they're linked to onto my laptop to put them
in the static folder.

The competition page has a red underlined text portion in the
illustration section that we'll need to discuss.  It includes all other
changes Debra made as regular text.