game page 90%
I had to redo some of the tables which was for some reason super
difficult. I ended up just starting a new DB over and copying all the
files over (which is why i made this a "fork" just incase). Right now,
each scenario will only have one problem and is only formatted correctly
for 3 solutions. the game index page also got a little weird, as it
shows the same scenario for as many solutions are available.