Themes Ready for Testing
So the store page is ready for testing, along with themes.  User pics
are NOT going to be tested (no default user picture is set up for the
player at the moment).  The actual purchasing logic for tokens is
figured out and commented, but not implemented for testing since we
don't want people to sit for 50+clicks just to switch themes.

As of right now, clicking on the "Purchase" button next to a theme will
change your theme and stick with you throughout the site.  Except for
the victory page, for some reason the theme won't stay with that one.

I also fixed the link colors (we can change them to actual buttons next
semester) so that the link color matches your theme text color instead
of staying blue/purple all the time.