accidently submitted before I commented.
As I meant to say in last push, I still can't figure out how to get
images to upload in. I get some weird error when about MulitKeyError
when trying to save the image in the DB. I tried just linking the images
manually after copying them into the folder they should be in, and no
luck there either. I added a bunch of stuff that may be on the right
track or just over complicating things.