A lot of the final binder documentation
A lot of it is done, but there are still things in it that I didn't know
what to put in for it. It also will probably need some screenshots of
the game. Also, it looked really good as far as formatting goes in
MSWord, and it looked pretty good in LibreOffice, except that the code
went from page 33-49, instead of 33-48 in Word.