Pictures work!
I finally figured out what was wrong, a '/' was in the causing all sorts
of problems. But now you can upload files from the admin page.

I had to do some migrations to the db, so it might have lost some
usernames and points, but I don't think anything major was missing.

I haven't uploaded to the server yet.