Started a user profile.
It doesn't quite work yet. I added in a couple items into the store and
linked them as purchases to player1. They don't show up correctly in the
profile page though. I did add a link to the page in the home page.
Within the profile page, there is a link under the avatar picture to
another page to delete or reset the account. I didn't set up the player
deletion, and didn't test the player reset.