Changed purchases
When a Player makes an account, the player's account will be linked to
all available purchases with a boolean of false (doesn't own), with the
exception of the two default purchases (default theme and picture).

I believe that Tim's version of the store_themes page is still there,
works on the live server, but not correctly on my machine. But anyhow,
the only change from I kept was in views, the page will only receive
item_list filled with the category "Themes," though I didn't remove the
search logic from the store_themes page.