Player groups added
I added a new table in for PLAYERGROUPS. After looking into the default
groups feature in the admin page for a while, I'm still not sure how
that would be beneficial to anything. Anyhow, admin's can easily link
players or a group of players to a scenario. A player can be selected in
both player_list and group_list, or just one for the the scenario to
correctly show up.