Fixes?
I started a fresh project and just copied all the code over, then
re-entered all the stuff in the DB. Found a couple missing lines of code
when a user makes an account and fixed that.

To my knowledge, registering a player works fine and the scenarios show
up. Something went wrong on the game page though and the popups aren't
working correctly.