Different type of popup
I figured out how to do an actual popup and thought maybe it could work
by having the game page load the next page right away after selecting an
answer, but having the popup open with the tokens earned. This is as
apposed to how it was before, where the player stays on the current
problem, until the popup is closed. However, I can't figure out how to
get it to work either. The popups can only work as a <a> but forms don't
accept <a>'s, just <inputs>.

I commented out the original way in case the new way doesn't work at
all.