Last Minute "Fixes"
Game page: The popup now has a color scheme matching the users theme.
On my machine, the gradient changed right below the last line of text so
there was not text/background color conflicts.

Store page: Changed purchase buttons to be token picture, and changed
some text inconsistencies.  For example, it used to say to click the
"apply" button to change themes, and now it has a picture of the
paintbrush image there instead.

Profile page: Fixed the issue where the theme previews weren't
displaying an actual color preview of the theme.